Cleanup our cache mess
Our ID-based caches turned out to be messy, were problematic when implementing disappearing messages, and they were not contained in a single class. This commit simplifies things: The one and only storage of messages in the app's memory is now the VirtualizedMessagesCollection, which maintains DbIdToMessageMap so that we get a reference to a SignalMessage's Message.

closes #74